image_analysis/requirements.txt
google-adk
google-cloud-aiplatform
opentelemetry-instrumentation-google-genai
python-dotenv
requests
Pillow
